requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   groq
    #   httpx
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==6.2.6
    # via streamlit
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
colorama==0.4.6
    # via click
distro==1.9.0
    # via groq
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
groq==0.37.1
    # via langchain-groq
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   groq
    #   langgraph-sdk
    #   langsmith
    #   ollama
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
langchain==1.2.9
    # via -r requirements.in
langchain-core==1.2.9
    # via
    #   langchain
    #   langchain-groq
    #   langchain-ollama
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-groq==1.1.2
    # via -r requirements.in
langchain-ollama==1.0.1
    # via -r requirements.in
langgraph==1.0.8
    # via langchain
langgraph-checkpoint==4.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.7
    # via langgraph
langgraph-sdk==0.3.4
    # via langgraph
langsmith==0.6.9
    # via langchain-core
markupsafe==3.0.3
    # via jinja2
narwhals==2.16.0
    # via altair
numpy==2.4.2
    # via
    #   pandas
    #   pydeck
    #   streamlit
ollama==0.6.1
    # via langchain-ollama
orjson==3.11.7
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.0
    # via
    #   altair
    #   langchain-core
    #   langsmith
    #   streamlit
pandas==2.3.3
    # via
    #   -r requirements.in
    #   streamlit
pillow==12.1.0
    # via streamlit
protobuf==6.33.5
    # via streamlit
pyarrow==23.0.0
    # via streamlit
pydantic==2.12.5
    # via
    #   groq
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
    #   ollama
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via langchain-core
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   langsmith
    #   requests-toolbelt
    #   streamlit
requests-toolbelt==1.0.0
    # via langsmith
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via groq
streamlit==1.54.0
    # via -r requirements.in
tenacity==9.1.4
    # via
    #   langchain-core
    #   streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   anyio
    #   groq
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
uuid-utils==0.14.0
    # via
    #   langchain-core
    #   langsmith
watchdog==6.0.0
    # via streamlit
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
zstandard==0.25.0
    # via langsmith
